This commit was manufactured by cvs2svn to create tag 'v3-02-01'.
git-svn-id: http://root.cern.ch/svn/root/tags/v3-02-01@3894 27541ba8-7e3a-0410-8455-c3a389f83636

+++ build/package/lib/makedebscr.sh
@@ -1,19 +1,19 @@
 #!/bin/sh -e 
 #
-# $Id: makedebscr.sh,v 1.1 2001/04/23 14:11:47 rdm Exp $
+# $Id$
 #
 # Writes a script files to debian/root-<pkg>.<scr> 
 #
+. build/package/lib/common.sh debian
+
+if [ $# -lt 2 ] ; then 
+    echo "$0: I need a package and script name - giving up"
+    exit 2
+fi
 
 # save package name in logical variable 
-tgtdir=$1 ; shift
-debdir=$1 ; shift
-cmndir=$1 ; shift 
-prefix=$1 ; shift
-etcdir=$1 ; shift 
-docdir=$1 ; shift 
-pkg=$1    ; shift
-scr=$1    ; shift
+pkg=$1
+scr=$2
 
 # make sure we get a fresh file 
 rm -f ${tgtdir}/{pkg}.${scr}
@@ -21,11 +21,7 @@ rm -f ${tgtdir}/{pkg}.${scr}
 # test to see if full file exist, if so, cat it to standard out. 
 if [ -f ${debdir}/${pkg}.${scr} ] 
 then 
-    sed -e "s,@prefix@,/${prefix},g" \
-	-e "s,@etcdir@,/${etcdir},g" \
-	-e "s,@docdir@,/${docdir},g" \
-	< ${debdir}/${pkg}.${scr} > ${tgtdir}/${pkg}.${scr}
-
+    cp ${debdir}/${pkg}.${scr} ${tgtdir}/${pkg}.${scr}
 # if a skeleton and body exist, we insert the body into the skeleton 
 # and cat it to standard out 
 elif [ -f ${cmndir}/${pkg}.${scr} ] && [ -f ${debdir}/${pkg}.${scr}.in ] 
@@ -35,24 +31,16 @@ then
     csplit -q -f ${cmndir}/tmp. \
 	-k ${debdir}/${pkg}.${scr}.in "/@${scr}@/"
 
-    # Then output the skeleton full file to temporary
+    # Then output the full file 
     cat ${cmndir}/tmp.00 \
 	${cmndir}/${pkg}.${scr} \
 	${cmndir}/tmp.01 | \
-	sed "/${scr}/d"  > ${cmndir}/tmp.02 
-
-    # Do expansion on file 
-    sed -e "s,@prefix@,/${prefix},g" \
-	-e "s,@etcdir@,/${etcdir},g" \
-	< ${cmndir}/tmp.02 > ${tgtdir}/${pkg}.${scr}
+	sed "/${scr}/d"  > ${tgtdir}/${pkg}.${scr}
 
     # Clean up
-    rm -f ${cmndir}/tmp.00 ${cmndir}/tmp.01 ${cmndir}/tmp.02
+    rm -f ${cmndir}/tmp.00 ${cmndir}/tmp.01
 fi
 
 #
-# $Log: makedebscr.sh,v $
-# Revision 1.1  2001/04/23 14:11:47  rdm
-# part of the debian and redhat build system.
-#
+# $Log$
 #
